refactor file download logic
