+++ packages/core/src/file.ts
@@ -3,8 +3,13 @@ import { readText } from "./fs"
 import { lookupMime } from "./mime"
 import { isBinaryMimeType } from "./parser"
 import { PDFPagesToString, PDFTryParse } from "./pdf"
+import { createFetch } from "./fetch"
+import { fileTypeFromBuffer } from "file-type"
+import { toBase64 } from "./util"
+import { host } from "./host"
+import { MarkdownTrace } from "./trace"
 
-export async function resolveFileContent(file: LinkedFile) {
+export async function resolveFileContent(file: LinkedFile, options?: { binary?: boolean }) {
     const { filename } = file
     if (file.content) return file
 
@@ -15,9 +20,28 @@ export async function resolveFileContent(file: LinkedFile) {
         file.content = await DOCXTryParse(filename)
     } else {
         const mime = lookupMime(filename)
-        const binary = isBinaryMimeType(mime)
-        file.content = binary ? undefined : await readText(filename)
+        const isBinary = isBinaryMimeType(mime)
+        if (!isBinary)
+            file.content = await readText(filename)
     }
-
     return file
+}
+
+export async function resolveFileDataUri(file: LinkedFile) {
+    let bytes: Uint8Array
+    if (/^https?:\/\//i.test(file.filename)) {
+        const fetch = await createFetch()
+        const resp = await fetch(file.filename)
+        const buffer = await resp.arrayBuffer()
+        bytes = new Uint8Array(buffer)
+    } else {
+        bytes = new Uint8Array(
+            await host.readFile(file.filename)
+        )
+    }
+    const mime = (await fileTypeFromBuffer(bytes))?.mime
+    if (!mime)
+        return undefined
+    const b64 = toBase64(bytes)
+    return `data:${mime};base64,${b64}`
 }
